requirements.txt
Flask==3.0.0
Flask-Babel==3.1.0